switched to santuario for c14n11 without comments
Removed dependency on axiom-c14n, which was formerly used solely
for xml canonicalization. Santuario does the same thing, but supports
a newer version of XML Canonicalization (1.1) and doesn't bring in
half a dozen additional dependencies like axiom-c14n does.

Switched to using XML Canonicalization 1.1 (without comments) because
it addresses several ugly namespace issues with the 1.0 version of
the spec.  In particular, it drops unnecessary namespace declarations
and moves common declarations up the tree, resulting in a more readable,
concise document.